Make next project text more responsive
Text should stick to the arrow as this did not happen on the very smallest of screens. Also reduce text size and margins on very small screens.
Center layout icons and text so that it looks good on small screens where they are closer together and differences in alignment are noticable.